flower_shop/requirements.txt
Django==3.2
python-telegram-bot==13.7
